hardware working :D

+++ code_2.1/Header/analog.h
@@ -7,7 +7,7 @@
 #define SAMPLE_COUNT 64
 
 // Global variables 
-extern volatile uint16_t sample_buffer[SAMPLE_COUNT];  // Buffer to store samples
+extern volatile uint32_t sample_buffer[SAMPLE_COUNT];  // Buffer to store samples
 extern volatile uint8_t sample_index;  // Index for buffer
 extern volatile bool bufferFullFlag;
